Implemented Command Pattern for player movement
Changed the player movement so that it works via the command pattern and unity input system now. Useful! Can now bind any key to the different move commands using the Actions file. Commands are stored in an array, could be used to do stuff like Undo actions. Maybe a rewind feature in the future?